requirements.txt
matplotlib==3.6.2
natsort==8.2.0
numpy==1.23.4
opencv-python==4.6.0.66
pandas==1.5.1
Pillow==9.3.0
PyYAML==6.0
requests==2.28.1
scipy==1.9.3
seaborn==0.12.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.1
PyQt5==5.15.7